commit project

+++ src/main/java/messageboard/service/Impl/BoardServiceImpl.java
@@ -1,10 +1,13 @@
 package messageboard.service.Impl;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import messageboard.Dto.BoardDto;
 import messageboard.entity.Board;
+import messageboard.entity.Member;
 import messageboard.repository.BoardRepository;
 import messageboard.service.BoardService;
+import messageboard.service.MemberService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -12,27 +15,34 @@ import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityNotFoundException;
 import java.time.LocalDateTime;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @Service
 @Transactional
+@Slf4j
 @RequiredArgsConstructor
 public class BoardServiceImpl implements BoardService {
 
     private final BoardRepository boardRepository;
+    private final MemberService memberService;
 
 
     @Override
     public Board save(BoardDto boardDto) {
 
+        String username = boardDto.getMemberDto().getUsername();
+        Member byUsername = memberService.findByUsername(username);
         Board build = Board.builder()
                 .title(boardDto.getTitle())
                 .dateTime(LocalDateTime.now())
-                .writer(boardDto.getWriter())
+                .writer(byUsername.getUsername())
                 .password(boardDto.getPassword())
                 .content(boardDto.getContent())
                 .count(0)
+                .member(byUsername)
                 .views(0)
                 .build();
 
@@ -56,15 +66,27 @@ public class BoardServiceImpl implements BoardService {
         return boardRepository.findAll(pageable);
     }
 
+    /**
+     * true 반환시 삭제 성공 false 반환시 등록된 사용자만 삭제가능
+     * @param boardId
+     * @param memberUsername
+     * @return
+     */
     @Override
-    public void deleteBoard(Long boardId) {
+    public boolean deleteBoard(Long boardId,String memberUsername) {
+        Member byUsername = memberService.findByUsername(memberUsername);
         Optional<Board> boardOptional = boardRepository.findById(boardId);
         if (boardOptional.isPresent()) {
-            boardRepository.deleteById(boardId);
+            String writer = boardOptional.get().getWriter();
+            if (byUsername!=null && writer.equals(memberUsername))
+            {
+                boardRepository.deleteById(boardId);
+                return true;
+            }
         } else {
             throw new EntityNotFoundException("게시물이 존재하지 않습니다. ID: " + boardId);
         }
-
+        return false;
     }
 
     @Override
@@ -93,14 +115,18 @@ public class BoardServiceImpl implements BoardService {
         return boardRepository.save(board);
     }
 
-    public boolean passwordVerify(Long boardId,String password){
+    public Integer passwordVerify(Long boardId, String password,String username){
         Board byBoardId = findByBoardId(boardId);
         String boardPassword = byBoardId.getPassword();
+        Member memberUsername = memberService.findByUsername(username);
 
-        if (password.equals(boardPassword)){
-            return true;
+        if (boardPassword.equals(password)) {
+            if (memberUsername == null||!(byBoardId.getWriter().equals(username))) {
+                return 2;  // 비밀번호는 o ,사용자 인증은 x
+            }
+            return 1;   //성공시 1로반환
         }
-        return false;
+        return 0;
     }
 
 }
